requirements.txt
mediapipe==0.10.8
opencv-python==4.8.1.78
mysql-connector-python==8.0.33
flask==3.0.0
pyautogui==0.9.54
numpy==1.24.3
python-dotenv==1.0.0
protobuf==3.20.3 